requirements.txt
# Зависимости для Telegram-бота комплектовщика команд

# Telegram Bot API
aiogram>=3.0.0

# Работа с переменными окружения
python-dotenv>=1.0.0

# Веб-сервер для webhook (опционально)
aiohttp>=3.8.0

# Системные утилиты для управления процессами
psutil>=5.9.0

# Тестирование
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Линтер (опционально)
ruff>=0.1.0